Added alignas 16 for Uniform buffer object's variables mat4

+++ Project3/Engine/Model.cpp
@@ -179,7 +179,7 @@ namespace Engine {
 		}
 	}
 
-	void Model::updateUniformBuffer(uint32_t currentImage, VkExtent2D Extent) {
+	void Model::updateUniformBuffer(size_t currentImage, VkExtent2D Extent) {
 		static auto startTime = std::chrono::high_resolution_clock::now();
 
 		auto currentTime = std::chrono::high_resolution_clock::now();
@@ -189,6 +189,7 @@ namespace Engine {
 		ubo.model = glm::rotate(glm::mat4(1.0f), time * glm::radians(90.0f), glm::vec3(0.0f, 0.0f, 1.0f));
 		ubo.view = glm::lookAt(glm::vec3(2.0f, 2.0f, 2.0f), glm::vec3(0.0f, 0.0f, 0.0f), glm::vec3(0.0f, 0.0f, 1.0f));
 		ubo.proj = glm::perspective(glm::radians(45.0f), static_cast<float>(Extent.width) / static_cast<float>(Extent.height), 0.1f, 10.0f);
+
 
 		ubo.proj[1][1] *= -1;
 		memcpy(uniformBufferMapped[currentImage], &ubo, sizeof(ubo));
